requirements.txt
python-telegram-bot==21.6
aiohttp
python-dotenv
pandas
matplotlib
